Added option to change the width of the stroke of the nodes.
git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1805 267469b6-68fb-4578-8a4b-5844dc4013a7

+++ source/net/sourceforge/toscanaj/view/diagram/LineView.java
@@ -102,8 +102,8 @@ public class LineView extends CanvasItem {
 			extentRatio = (double)lowerExtent / (double)upperExtent;
 		}
 
-		int defaultLineWidth = diagramSchema.getDefaultLineWidth();
-		int selectionLineWidth = diagramSchema.getSelectionLineWidth();
+		float defaultLineWidth = diagramSchema.getDefaultLineWidth();
+		float selectionLineWidth = diagramSchema.getSelectionLineWidth();
 		Color lineColor = null;
 		float strokeWidth = defaultLineWidth;
 		if (this.diagramLine.getFromNode().getY() > this.diagramLine.getToNode().getY()) {
